# api/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Movie Hub
  description: "A Movie Hub API."
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: "Local server"
security:
  - bearerAuth: []
components:
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    movieId:
      name: movieId
      in: path
      description: "The movie ID."
      schema:
        type: string
        format: uuid
        pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89AB][0-9a-f]{3}-[0-9a-f]{12}$"
        maxLength: 255
      required: true
      example: 6c42ccf6-f635-48f8-9af5-4c80e6812dbf
    page:
      name: page
      in: query
      description: "The page number for pagination."
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 10000
        default: 0
        example: 0
    limit:
      name: limit
      in: query
      description: "The maximum number of items to return per page."
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 50
        default: 10
        example: 10
    sort:
      name: sort
      in: query
      description: |
        A comma-separated list of sort keys.
        Valid keys are 'name', 'releaseDate', 'duration'.
        Each key can be sorted by either: asc (ascending order) or desc (descending order).
        Example usage: ?sort=name.desc,duration.asc.
      schema:
        type: string
        pattern: "^(name|releaseDate|duration)\\.(asc|desc)(,(name|releaseDate|duration)\\.(asc|desc))*$"
        maxLength: 100
        default: name.asc
        example: name.asc
    releaseYear:
      name: releaseYear
      in: query
      description: |
        Query parameter for filtering movies based on their release year. 
        Specify a four-digit year to retrieve movies released in that year (e.g., "2024").
        To retrieve movies from all years, use the value "all".
      schema:
        type: string
        pattern: "^(all|(\\d{4}(,\\d{4})*))$"
        minLength: 3
        maxLength: 300
        default: all
        example: 2024
    genre:
      name: genre
      in: query
      description: |
        Query parameter for filtering movies based on their genres. 
        You can specify one or more genres as a comma-separated list (e.g., "Thriller,Romance").
        To retrieve all movies regardless of genre, use the value "all".
      schema:
        type: string
        pattern: "^([a-zA-Z]+(,[a-zA-Z]+)*)?$"
        minLength: 1
        maxLength: 300
        example: Thriller
        default: all
    country:
      name: country
      in: query
      description: |
        Query parameter for filtering movies based on their countries. 
        You can specify one or more countries as a comma-separated list (e.g., "South Korea,Japan").
        To retrieve all movies regardless of country, use the value "all".
      schema:
        type: string
        pattern: "^([a-zA-Z ]+(,[a-zA-Z ]+)*)?$"
        minLength: 1
        maxLength: 300
        example: South Korea
        default: all
    keyword:
      name: keyword
      in: query
      description: |
        A query parameter used to filter movies based on a specific keyword.
        This parameter is useful for searching movies by name.
      schema:
        type: string
        pattern: "^[A-Za-z0-9.,'\"&()\\-: ]*$"
        default: ""
        maxLength: 100
...
    commentLimit:
      name: limit
      in: query
      description: "The maximum number of items to return per page."
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 5
        default: 5
        example: 5
    commentSort:
      name: sort
      in: query
      description: |
        Only valid sort key is 'createdAt'.
        It can be sorted by either: asc (ascending order) or desc (descending order).
        Example usage: ?sort=createdAt.desc.
      schema:
        type: string
        pattern: "^createdAt\\.(asc|desc)$"
        default: createdAt.desc
        example: createdAt.desc
        maxLength: 200
...
    genreId:
      name: genreId
      in: path
      required: true
      schema:
        type: string
        format: uuid
        pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89AB][0-9a-f]{3}-[0-9a-f]{12}$"
        maxLength: 255
      description: "The genre ID."
      example: 6b0de5bc-8e6c-4d4c-b13e-49450d7480a7
...
  responses:
    404MovieNotFound:
      description: Movie not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    404GenreNotFound:
      description: Genre not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    401Unauthorized:
      description: Either no access token was provided or the provided access token is invalid or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    403Forbidden:
      description: The user does not have permission to access this resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Default:
      description: Unexpected error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
...
    GenreName:
      type: string
      example: Horror
      description: |
        The genre or category of the movie.
        This field represents the type of movie, such as Horror or Comedy, and is limited to 15 characters.
      pattern: "^[A-Za-z\\- ]+$"
      minLength: 1
      maxLength: 15
      nullable: false
...
    MovieUserRating:
      type: object
      additionalProperties: false
      description: "The authenticated user's rating of the movie."
      required:
        - rating
      properties:
        rating:
          type: number
          format: double
          description: "The authenticated user's rating (value) of the movie."
          example: 8.0
          minimum: 0
          maximum: 10
          default: 0.0
...
    MoviePage:
      type: object
      description: "Represents a page of movies, including pagination and sorting information."
      additionalProperties: false
      required:
        - content
        - pageable
        - last
        - totalElements
        - totalPages
        - first
        - size
        - number
        - sort
        - numberOfElements
        - empty
      properties:
        content:
          type: array
          description: "List of movies."
          maxItems: 50
          items:
            $ref: '#/components/schemas/MoviePreviewResponse'
        pageable:
          $ref: '#/components/schemas/PageableDTO'
        last:
          type: boolean
          description: |
            A boolean value indicating whether this is the last page of data.
            If `true`, it is the last page; otherwise, it's not the last page.
          example: false
        totalElements:
          type: integer
          format: int64
          description: "The total number of items."
          example: 4
        totalPages:
          type: integer
          description: "The total number of pages needed to display all the data based on the page size."
          example: 2
        first:
          type: boolean
          description: |
            A boolean value indicating whether this is the first page of data.
            If `true`, it is the first page; otherwise, it's not the first page.
          example: true
        size:
          type: integer
          description: "The maximum number of items contained on each page."
          example: 2
        number:
          type: integer
          description: "The page number for the current page."
          example: 0
        sort:
          $ref: '#/components/schemas/SortDTO'
        numberOfElements:
          type: integer
          description: "The number of items on the current page."
          example: 2
        empty:
          type: boolean
          description: |
            A boolean value indicating whether the current page is empty.
            If `true`, the page has no items; otherwise, it has content.
          example: false
...
    AddMovieRequest:
      type: object
      additionalProperties: false
      description: |
        Schema for adding a new movie to the database.
        It requires all movie details to be provided.
      required:
        - name
        - filename
        - releaseDate
        - duration
        - description
        - posterUrl
        - trailerUrl
        - cast
        - productionCompanies
        - genres
        - countries
      properties:
        name:
          $ref: '#/components/schemas/MovieName'
        filename:
          $ref: '#/components/schemas/MovieFilename'
        releaseDate:
          $ref: '#/components/schemas/MovieReleaseDate'
        duration:
          $ref: '#/components/schemas/MovieDuration'
        description:
          $ref: '#/components/schemas/MovieDescription'
        posterUrl:
          $ref: '#/components/schemas/MoviePosterUrl'
        trailerUrl:
          $ref: '#/components/schemas/MovieTrailerUrl'
        director:
          $ref: '#/components/schemas/PersonName'
        cast:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/MovieCastRequest'
        productionCompanies:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/ProductionCompanyName'
        genres:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/GenreName'
        countries:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/CountryName'
    UpdateMovieRequest:
      type: object
      additionalProperties: false
      description: |
        Schema for updating an existing movie's details.
        It allows partial updates to movie information based on the existing schema.
      properties:
        name:
          $ref: '#/components/schemas/MovieName'
        filename:
          $ref: '#/components/schemas/MovieFilename'
        releaseDate:
          $ref: '#/components/schemas/MovieReleaseDate'
        duration:
          $ref: '#/components/schemas/MovieDuration'
        description:
          $ref: '#/components/schemas/MovieDescription'
        posterUrl:
          $ref: '#/components/schemas/MoviePosterUrl'
        trailerUrl:
          $ref: '#/components/schemas/MovieTrailerUrl'
        director:
          $ref: '#/components/schemas/PersonName'
        cast:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/MovieCastRequest'
        productionCompanies:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/ProductionCompanyName'
        genres:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/GenreName'
        countries:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/CountryName'
    AddCommentRequest:
      type: object
      description: Schema for the requst body when creating a comment.
      additionalProperties: false
      required:
        - text
      properties:
        text:
          $ref: '#/components/schemas/Text'
...
    MovieDetailsResponse:
      type: object
      description: "Offers comprehensive details about a movie."
      additionalProperties: false
      required:
        - id
        - name
        - filename
        - duration
        - description
        - rating
        - reviewCount
        - posterUrl
        - trailerUrl
        - director
        - cast
        - productionCompanies
        - genres
        - countries
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          $ref: '#/components/schemas/MovieName'
        filename:
          $ref: '#/components/schemas/MovieFilename'
        releaseDate:
          $ref: '#/components/schemas/MovieReleaseDate'
        duration:
          $ref: '#/components/schemas/MovieDuration'
        description:
          $ref: '#/components/schemas/MovieDescription'
        rating:
          $ref: '#/components/schemas/MovieRating'
        reviewCount:
          $ref: '#/components/schemas/MovieReviewCount'
        posterUrl:
          $ref: '#/components/schemas/MoviePosterUrl'
        trailerUrl:
          $ref: '#/components/schemas/MovieTrailerUrl'
        director:
          $ref: '#/components/schemas/DirectorResponse'
        cast:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/MovieCastDetailsResponse'
        productionCompanies:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/ProductionCompanyDetailsResponse'
        genres:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/GenreDetailsResponse'
        countries:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/CountryDetailsResponse'
...
    GenreDetailsResponse:
      type: object
      description: "Provides details about a genre."
      additionalProperties: false
      required:
        - id
        - name
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          $ref: '#/components/schemas/GenreName'
...
    GenreListResponse:
      type: object
      description: "Provides a list of all genres available in the system."
      additionalProperties: false
      required:
        - genres
      properties:
        genres:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/GenreDetailsResponse'
          example:
            - id: 21c9c051-0bf4-4e2b-8c96-7c19895ce72f
              name: Action
            - id: 3d9e9db7-6544-4613-998e-88cdd6334ce1
              name: Horror
...
    YearListResponse:
      type: object
      description: "Provides a list of all distinct years of the movies available in the system."
      additionalProperties: false
      required:
        - years
      properties:
        years:
          type: array
          maxItems: 200
          description: "List of years."
          items:
            type: integer
            format: int32
            minimum: 1888
            maximum: 9999
            example: 2024
...
paths:
  /movies:
    post:
      description: Create a new movie record in the database with the specified details.
      summary: Add movie
      operationId: addMovie
      security:
        - bearerAuth: []
      tags:
        - movies
      requestBody:
        description: A movie to be added to the database.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMovieRequest'
      responses:
        '201':
          description: Successfully added movie to the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetailsResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        'default':
          $ref: '#/components/responses/Default'
    get:
      description: |
        Retrieve a list of movies in the system.
        The results can be customized with various filters, sorting options, and pagination to refine the list based on specific preferences.
      summary: List movies
      operationId: getMovies
      tags:
        - movies
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/sort'
      responses:
        '200':
          description: Successfully retrieved a list of movies in the system.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoviePage'
        'default':
          $ref: '#/components/responses/Default'
  /movies/filter:
    get:
      description: |
        Filter movies using various criteria to narrow down the search results based on specific preferences.
      summary: Filter movies
      operationId: filterMovies
      tags:
        - movies
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/sort'
        - $ref: '#/components/parameters/releaseYear'
        - $ref: '#/components/parameters/genre'
        - $ref: '#/components/parameters/country'
      responses:
        '200':
          description: Successfully retrieved a list of filtered movies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoviePage'
        'default':
          $ref: '#/components/responses/Default'
  /movies/search:
    get:
      description: |
        Search for movies using a keyword.
      summary: Search movies
      operationId: searchMovies
      tags:
        - movies
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/sort'
        - $ref: '#/components/parameters/keyword'
      responses:
        '200':
          description: Successfully retrieved a list of movies matching the search keyword.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoviePage'
        'default':
          $ref: '#/components/responses/Default'
  /movies/years:
    get:
      description: |
        Fetch distinct years of all the movies in the system.
      summary: List years
      operationId: getYears
      tags:
        - movies
      responses:
        '200':
          description: Successfully retrieved a list of all distinct years of the movies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/YearListResponse'
        'default':
          $ref: '#/components/responses/Default'
  /movies/{movieId}:
    get:
      description: Retrieve detailed information about a movie with the specified ID.
      summary: Get movie details
      operationId: getMovie
      tags:
        - movies
      parameters:
        - $ref: '#/components/parameters/movieId'
      responses:
        '200':
          description: Successfully retrieved detailed information about the movie with the specified ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetailsResponse'
        '404':
          $ref: '#/components/responses/404MovieNotFound'
        'default':
          $ref: '#/components/responses/Default'
    put:
      description: |
        Update the details of an existing movie.
        All movie details will be replaced with the new details provided.
        For example, all the cast members provided will be saved as the new cast of the movie.
      summary: Update movie details
      operationId: updateMovie
      security:
        - bearerAuth: []
      tags:
        - movies
      parameters:
        - $ref: '#/components/parameters/movieId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMovieRequest'
      responses:
        '200':
          description: Successfully updated the movie details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetailsResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404MovieNotFound'
        'default':
          $ref: '#/components/responses/Default'
    delete:
      description: Delete a movie with the specified ID from the database.
      summary: Delete movie
      operationId: deleteMovie
      security:
        - bearerAuth: []
      tags:
        - movies
      parameters:
        - $ref: '#/components/parameters/movieId'
      responses:
        '204':
          description: Successfully deleted the movie from the database.
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404MovieNotFound'
        'default':
          $ref: '#/components/responses/Default'
  /movies/{movieId}/ratings:
    post:
      description: Add a rating for the specified movie.
      summary: Add movie rating
      operationId: addRating
      security:
        - bearerAuth: []
      tags:
        - movies
      parameters:
        - $ref: '#/components/parameters/movieId'
      requestBody:
        description: The rating of the movie.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieUserRating'
      responses:
        '204':
          description: Successfully added the movie rating.
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404MovieNotFound'
        'default':
          $ref: '#/components/responses/Default'
  /movies/{movieId}/ratings/me:
    get:
      description: Get the current user's rating for a specified movie.
      summary: Get user's rating
      operationId: getUserRating
      security:
        - bearerAuth: []
      tags:
        - movies
      parameters:
        - $ref: '#/components/parameters/movieId'
      responses:
        '200':
          description: Successfully retrieved the user's movie rating.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieUserRating'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404MovieNotFound'
        'default':
          $ref: '#/components/responses/Default'
  /movies/{movieId}/comments:
    post:
      description: Add a comment for the specified movie.
      summary: Add comment
      operationId: addComment
      security:
        - bearerAuth: []
      tags:
        - movies
      parameters:
        - $ref: '#/components/parameters/movieId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCommentRequest'
      responses:
        '201':
          description: Sucessfully added a comment for the specified movie.
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404MovieNotFound'
        'default':
          $ref: '#/components/responses/Default'

    get:
      description: Retrieve a list of comments of the specified movie.
      summary: List comments
      operationId: getComments
      tags:
        - movies
      parameters:
        - $ref: '#/components/parameters/movieId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/commentLimit'
        - $ref: '#/components/parameters/commentSort'
      responses:
        '200':
          description: Successfully retrieved a list of comments of the specified movie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentInfoPage'
        '404':
          $ref: '#/components/responses/404MovieNotFound'
        'default':
          $ref: '#/components/responses/Default'
...
  /genres:
    post:
      description: |
        Create a new genre record in the database with the specified details.
        This operation adds a genre to the system, enabling it to be used in association with movies.
      summary: Add genre
      operationId: addGenre
      security:
        - bearerAuth: []
      tags:
        - genres
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  $ref: '#/components/schemas/GenreName'
      responses:
        '201':
          description: Successfully created a new genre record in the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreDetailsResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '409':
          description: The genre already exists in the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        'default':
          $ref: '#/components/responses/Default'
    get:
      description: Retrieve a list of genres available in the system.
      summary: List genres
      operationId: getGenres
      tags:
        - genres
      responses:
        '200':
          description: Successfully retrieved the list of genres available in the system.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreListResponse'
        'default':
          $ref: '#/components/responses/Default'
  /genres/{genreId}:
    get:
      description: Retrieve detailed information about a specific genre.
      summary: Get genre details
      operationId: getGenre
      tags:
        - genres
      parameters:
        - $ref: '#/components/parameters/genreId'
      responses:
        '200':
          description: Successfully retrieved the details of the specified genre.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreDetailsResponse'
        '404':
          $ref: '#/components/responses/404GenreNotFound'
        'default':
          $ref: '#/components/responses/Default'
    put:
      description: |
        Update the details of an existing genre.
        All genre's details will be replaced with the new details provided.
      summary: Update genre details
      operationId: updateGenre
      security:
        - bearerAuth: []
      tags:
        - genres
      parameters:
        - $ref: '#/components/parameters/genreId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  $ref: '#/components/schemas/GenreName'
      responses:
        '200':
          description: Successfully updated the details of the specified genre.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreDetailsResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404GenreNotFound'
        'default':
          $ref: '#/components/responses/Default'
  /genres/{genreId}/movies:
    get:
      description: Retrieve a list of movies that are categorized under the specified genre.
      summary: List movies with genre
      operationId: getMoviesWithGenre
      tags:
        - genres
      parameters:
        - $ref: '#/components/parameters/genreId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Successfully retrieved a list of movies that are categorized under the specified genre.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoviePage'
        '404':
          $ref: '#/components/responses/404GenreNotFound'
        'default':
          $ref: '#/components/responses/Default'
...